requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
attrs==23.1.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
distlib==0.3.7
exceptiongroup==1.1.3
filelock==3.12.4
flake8==6.1.0
identify==2.5.30
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
pytest==7.4.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.24.5
